livekit_agent/requirements.txt
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]>=1.0.0rc5
bithuman>=0.4.9
python-dotenv~=1.1
sounddevice~=0.5